refactor: split register-routes.ts into domain route modules (Issue #484)
- Extract Env and toApiRoutePath to routes/env.ts
- Create domain routers: auth, billing, rag, library, upload, assessment,
  onboarding, external-resources, telemetry, progress, notifications, chat
- Split campaigns into 9 sub-routers (core, planning, share, world-state,
  session-digests, entities, communities, graphrag, graph-rebuild)
- Add app module for SPA, assets, agents, 404 catch-all
- Thin routes/index.ts mounts all domain routers
- Remove register-routes.ts (1,415 lines)
- Update server.ts, billing.ts, test-env.ts, vitest.config

Made-with: Cursor
