+++ src/routes/campaigns/entities-routes.ts
@@ -1,0 +1,156 @@
+import type { Hono } from "hono";
+import type { RequestLogger } from "@/lib/logger";
+import { requireUserJwt } from "@/routes/auth";
+import {
+	handleCreateEntityRelationship,
+	handleDeleteEntityRelationship,
+	handleGetEntity,
+	handleGetEntityImportance,
+	handleGetEntityNeighbors,
+	handleGetEntityRelationships,
+	handleListEntities,
+	handleListPendingDeduplication,
+	handleListRelationshipTypes,
+	handleListTopEntitiesByImportance,
+	handleResolveDeduplicationEntry,
+	handleTestEntityExtractionFromR2,
+	handleTriggerEntityDeduplication,
+	handleTriggerEntityExtraction,
+	handleUpdateEntityImportance,
+} from "@/routes/entities";
+import type { Env } from "@/routes/env";
+import { toApiRoutePath } from "@/routes/env";
+import { API_CONFIG } from "@/shared-config";
+
+export function registerCampaignEntitiesRoutes(
+	app: Hono<{ Bindings: Env; Variables: { logger: RequestLogger } }>
+) {
+	app.get(
+		toApiRoutePath(API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.LIST(":campaignId")),
+		requireUserJwt,
+		handleListEntities
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.DETAILS(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleGetEntity
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.RELATIONSHIPS(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleGetEntityRelationships
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.GRAPH_NEIGHBORS(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleGetEntityNeighbors
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.RELATIONSHIP_TYPES(":campaignId")
+		),
+		requireUserJwt,
+		handleListRelationshipTypes
+	);
+	app.patch(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.IMPORTANCE(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleUpdateEntityImportance
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.IMPORTANCE(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleGetEntityImportance
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.IMPORTANCE_TOP(":campaignId")
+		),
+		requireUserJwt,
+		handleListTopEntitiesByImportance
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.RELATIONSHIPS(
+				":campaignId",
+				":entityId"
+			)
+		),
+		requireUserJwt,
+		handleCreateEntityRelationship
+	);
+	app.delete(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.RELATIONSHIP_DETAIL(
+				":campaignId",
+				":entityId",
+				":relationshipId"
+			)
+		),
+		requireUserJwt,
+		handleDeleteEntityRelationship
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.EXTRACT(":campaignId")
+		),
+		requireUserJwt,
+		handleTriggerEntityExtraction
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.DEDUPLICATE(":campaignId")
+		),
+		requireUserJwt,
+		handleTriggerEntityDeduplication
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.DEDUP_PENDING(":campaignId")
+		),
+		requireUserJwt,
+		handleListPendingDeduplication
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.DEDUP_RESOLVE(
+				":campaignId",
+				":entryId"
+			)
+		),
+		requireUserJwt,
+		handleResolveDeduplicationEntry
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.ENTITIES.TEST_EXTRACT_FROM_R2
+		),
+		requireUserJwt,
+		handleTestEntityExtractionFromR2
+	);
+}
