+++ src/routes/campaigns/session-digests-routes.ts
@@ -1,0 +1,150 @@
+import type { Hono } from "hono";
+import type { RequestLogger } from "@/lib/logger";
+import { requireUserJwt } from "@/routes/auth";
+import type { Env } from "@/routes/env";
+import { toApiRoutePath } from "@/routes/env";
+import {
+	handleCreateSessionDigestTemplate,
+	handleDeleteSessionDigestTemplate,
+	handleGetSessionDigestTemplate,
+	handleGetSessionDigestTemplates,
+	handleUpdateSessionDigestTemplate,
+} from "@/routes/session-digest-templates";
+import {
+	handleApproveDigest,
+	handleCreateSessionDigest,
+	handleDeleteSessionDigest,
+	handleGetSessionDigest,
+	handleGetSessionDigests,
+	handleRejectDigest,
+	handleSubmitDigestForReview,
+	handleUpdateSessionDigest,
+} from "@/routes/session-digests";
+import { API_CONFIG } from "@/shared-config";
+
+export function registerCampaignSessionDigestsRoutes(
+	app: Hono<{ Bindings: Env; Variables: { logger: RequestLogger } }>
+) {
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.BASE(":campaignId")
+		),
+		requireUserJwt,
+		handleCreateSessionDigest
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.BASE(":campaignId")
+		),
+		requireUserJwt,
+		handleGetSessionDigests
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.DETAILS(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleGetSessionDigest
+	);
+	app.put(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.DETAILS(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleUpdateSessionDigest
+	);
+	app.delete(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.DETAILS(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleDeleteSessionDigest
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.SUBMIT(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleSubmitDigestForReview
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.APPROVE(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleApproveDigest
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGESTS.REJECT(
+				":campaignId",
+				":digestId"
+			)
+		),
+		requireUserJwt,
+		handleRejectDigest
+	);
+	app.post(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGEST_TEMPLATES.BASE(
+				":campaignId"
+			)
+		),
+		requireUserJwt,
+		handleCreateSessionDigestTemplate
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGEST_TEMPLATES.BASE(
+				":campaignId"
+			)
+		),
+		requireUserJwt,
+		handleGetSessionDigestTemplates
+	);
+	app.get(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGEST_TEMPLATES.DETAILS(
+				":campaignId",
+				":templateId"
+			)
+		),
+		requireUserJwt,
+		handleGetSessionDigestTemplate
+	);
+	app.put(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGEST_TEMPLATES.DETAILS(
+				":campaignId",
+				":templateId"
+			)
+		),
+		requireUserJwt,
+		handleUpdateSessionDigestTemplate
+	);
+	app.delete(
+		toApiRoutePath(
+			API_CONFIG.ENDPOINTS.CAMPAIGNS.SESSION_DIGEST_TEMPLATES.DETAILS(
+				":campaignId",
+				":templateId"
+			)
+		),
+		requireUserJwt,
+		handleDeleteSessionDigestTemplate
+	);
+}
